requirements.txt
asgiref==3.2.10
backcall==0.2.0
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorlover==0.3.0
dash==1.17.0
dash-core-components==1.13.0
dash-daq==0.5.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
decorator==4.4.2
Django==3.1.1
django-ckeditor==6.0.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-plotly-dash==1.4.2
dpd-components==0.1.0
Flask==1.1.2
Flask-Compress==1.7.0
future==0.18.2
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jupyter-client==6.1.11
jupyter-core==4.7.1
lxml==4.6.1
MarkupSafe==1.1.1
mysqlclient==2.0.1
numpy==1.19.3
pandas==1.1.4
pandas-datareader==0.9.0
parso==0.8.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.12.0
plotly-express==0.4.1
prompt-toolkit==3.0.14
Pygments==2.7.4
python-dateutil==2.8.1
pytz==2020.1
pywin32==300
pyzmq==22.0.2
requests==2.24.0
retrying==1.3.3
scipy==1.5.3
six==1.15.0
sqlparse==0.3.1
statsmodels==0.12.1
tornado==6.1
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
